% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_tab.R
\name{adj_tab}
\alias{adj_tab}
\title{Extract adjusted 2-by-2 table from episensr object}
\usage{
adj_tab(x)
}
\arguments{
\item{x}{An object of class 'episensr'.}
}
\value{
A matrix.
}
\description{
Extract the adjusted 2-by-2 table from an episensr function, so that it can be re-used
into an other episensr function.
}
\examples{
# The data for this example come from:
# Stang A., Schmidt-Pokrzywniak A., Lehnert M., Parkin D.M., Ferlay J., Bornfeld N.
# et al.
# Population-based incidence estimates of uveal melanoma in Germany. Supplementing
# cancer registry data by case-control data.
# Eur J Cancer Prev 2006;15:165-70.
selection(matrix(c(136, 107, 297, 165),
dimnames = list(c("UM+", "UM-"), c("Mobile+", "Mobile-")),
nrow = 2, byrow = TRUE),
bias_parms = c(.94, .85, .64, .25))
}
