% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{callback_df}
\alias{callback_df}
\title{Given a contingency table (2-by-2) (as a data frame or matrix)
it will untabulate it by repeating each row by the number of times
it was repeated.}
\usage{
callback_df(df, num)
}
\arguments{
\item{matrix}{or data.frame to untable}

\item{vector}{of counts (of same length as \code{df})}
}
\description{
Given a contingency table (2-by-2) (as a data frame or matrix)
it will untabulate it by repeating each row by the number of times
it was repeated.
}
\keyword{internal}
