% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounders.emm.R
\name{confounders.emm}
\alias{confounders.emm}
\title{Sensitivity analysis to correct for unknown or unmeasured confounding with
effect modification}
\usage{
confounders.emm(case, exposed, type = c("RR", "OR", "RD"),
  bias_parms = NULL, alpha = 0.05, dec = 4, print = TRUE)
}
\arguments{
\item{case}{Outcome variable. If a variable, this variable is tabulated against.}

\item{exposed}{Exposure variable.}

\item{type}{Choice of implementation, with no effect measure modification for
ratio measures (relative risk -- RR; odds ratio -- OR) or difference measures
(risk difference -- RD).}

\item{bias_parms}{Numeric vector defining the 4 necessary bias parameters. This
vector has 4 elements, in the following order:
\enumerate{
\item the association between the confounder and the outcome among those who
were exposed,
\item the association between the confounder and the outcome among those who
were not exposed,
\item the prevalence of the confounder among the exposed, and
\item the prevalence of the confounder among the unexposed.
}}

\item{alpha}{Significance level.}

\item{dec}{Number of decimals in the printout.}

\item{print}{A logical scalar. Should the results be printed?}
}
\value{
A list with elements:
\item{obs.data}{The analysed 2 x 2 table from the observed data.}
\item{cfder.data}{The same table for Confounder +.}
\item{nocfder.data}{The same table for Confounder -.}
\item{obs.measures}{A table of relative risk with confidence intervals; Total,
for Confoundner +, and for Confounder -.}
\item{adj.measures}{A table of Standardized Morbidity Ratio and Mantel-Haenszel
estimates.}
\item{bias.parms}{Input bias parameters.}
}
\description{
Simple sensitivity analysis to correct for unknown or unmeasured confounding
with effect measure modification. Implementation for ratio measures (relative
risk -- RR, or odds ratio -- OR) and difference measures (risk difference -- RD).
}
\examples{
# The data for this example come from:
# Tyndall M.W., Ronald A.R., Agoki E., Malisa W., Bwayo J.J., Ndinya-Achola J.O.
# et al.
# Increased risk of infection with human immunodeficiency virus type 1 among
# uncircumcised men presenting with genital ulcer disease in Kenya.
# Clin Infect Dis 1996;23:449-53.
confounders.emm(matrix(c(105, 85, 527, 93),
dimnames = list(c("HIV+", "HIV-"), c("Circ+", "Circ-")),
nrow = 2, byrow = TRUE),
type = "RR",
bias_parms = c(.4, .7, .8, .05))
confounders.emm(matrix(c(105, 85, 527, 93),
dimnames = list(c("HIV+", "HIV-"), c("Circ+", "Circ-")),
nrow = 2, byrow = TRUE),
type = "OR",
bias_parms = c(.4, .7, .8, .05))
confounders.emm(matrix(c(105, 85, 527, 93),
dimnames = list(c("HIV+", "HIV-"), c("Circ+", "Circ-")),
nrow = 2, byrow = TRUE),
type = "RD",
bias_parms = c(-.6, -.3, .8, .05))
}
\references{
Lash, T.L., Fox, M.P, Fink, A.K., 2009 \emph{Applying Quantitative
Bias Analysis to Epidemiologic Data}, pp.59--78, Springer.
}

