% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounders.poly.R
\name{confounders.poly}
\alias{confounders.poly}
\title{Sensitivity analysis to correct for unknown or unmeasured polychotomous confounding without effect modification}
\usage{
confounders.poly(
  case,
  exposed,
  type = c("RR", "OR", "RD"),
  bias_parms = NULL,
  alpha = 0.05
)
}
\arguments{
\item{case}{Outcome variable. If a variable, this variable is tabulated against.}

\item{exposed}{Exposure variable.}

\item{type}{Choice of implementation, with no effect measure modification for
ratio measures (relative risk -- RR; odds ratio -- OR) or difference measures
(risk difference -- RD).}

\item{bias_parms}{Numeric vector defining the bias parameters. This vector
has 6 elements, in the following order:
\enumerate{
\item the association between the highest level confounder and the outcome,
\item the association between the mid-level confounder and the outcome,
\item the prevalence of the highest level confounder among the exposed (between 0 and 1),
\item the prevalence of the highest level confounder among the unexposed (between 0 and 1),
\item the prevalence of the mid-level confounder among the exposed (between 0 and 1), and
\item the prevalence of the mid-level confounder among the unexposed (between 0 and 1).
}}

\item{alpha}{Significance level.}
}
\value{
A list with elements:
\item{obs.data}{The analyzed 2 x 2 table from the observed data.}
\item{cfder1.data}{The same table for Mid-level Confounder +.}
\item{cfder2.data}{The same table for Highest-level Confounder +.}
\item{nocfder.data}{The same table for Confounder -.}
\item{obs.measures}{A table of relative risk with confidence intervals; Total
and by confounders.}
\item{adj.measures}{A table of Standardized Morbidity Ratio and Mantel-Haenszel
estimates.}
\item{bias.parms}{Input bias parameters.}
}
\description{
Simple sensitivity analysis to correct for unknown or unmeasured polychotomous
(3-level) confounding without effect modification. Implementation for ratio
measures (relative risk -- RR, or odds ratio -- OR) and difference measures
(risk difference -- RD).
}
\examples{
# The data for this example come from:
# Tyndall M.W., Ronald A.R., Agoki E., Malisa W., Bwayo J.J., Ndinya-Achola J.O.
# et al.
# Increased risk of infection with human immunodeficiency virus type 1 among
# uncircumcised men presenting with genital ulcer disease in Kenya.
# Clin Infect Dis 1996;23:449-53.
confounders.poly(matrix(c(105, 85, 527, 93),
dimnames = list(c("HIV+", "HIV-"), c("Circ+", "Circ-")),
nrow = 2, byrow = TRUE),
type = "RR",
bias_parms = c(.4, .8, .6, .05, .2, .2))

confounders.poly(matrix(c(105, 85, 527, 93),
dimnames = list(c("HIV+", "HIV-"), c("Circ+", "Circ-")),
nrow = 2, byrow = TRUE),
type = "OR",
bias_parms = c(.4, .8, .6, .05, .2, .2))

confounders.poly(matrix(c(105, 85, 527, 93),
dimnames = list(c("HIV+", "HIV-"), c("Circ+", "Circ-")),
nrow = 2, byrow = TRUE),
type = "RD",
bias_parms = c(-.4, -.2, .6, .05, .2, .2))
}
\references{
Lash, T.L., Fox, M.P, Fink, A.K., 2009 \emph{Applying Quantitative
Bias Analysis to Epidemiologic Data}, pp.59--78, Springer.
}
