% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episensr-package.R
\docType{package}
\name{episensr-package}
\alias{episensr-package}
\alias{episensr}
\title{\strong{episensr}: Basic Sensitivity Analysis for Epidemiological Results}
\description{
\strong{episensr} allows to do basic sensitivity analysis of epidemiological
results as described in \strong{Applying Quantitative Bias Analysis to
Epidemiological Data, 2nd ed.} by Matthew P. Fox, Richard F. MacLehose,
and Timothy L. Lash
}
\seealso{
Useful links:
\itemize{
  \item \url{https://codeberg.org/dhaine/episensr}
  \item \url{https://dhaine.codeberg.page/episensr/}
  \item Report bugs at \url{https://codeberg.org/dhaine/episensr/issues}
}

}
\author{
\strong{Maintainer}: Denis Haine \email{cheval@zaclys.net} (\href{https://orcid.org/0000-0002-6691-7335}{ORCID})

}
\keyword{internal}
