% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.episensr.booted.R
\name{print.episensr.booted}
\alias{print.episensr.booted}
\title{Print bootstraped confidence intervals}
\usage{
\method{print}{episensr.booted}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{An object of class 'episensr.booted'.}

\item{digits}{Minimal number of _significant_ digits, see 'print.default'.}

\item{...}{Other unused arguments.}
}
\value{
Print the confidence interval of the adjusted measures of association.
}
\description{
Print bootstraped confidence intervals for selection and misclassification bias functions.
}

