% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.mbias.R
\name{print.mbias}
\alias{print.mbias}
\title{Print association corrected for M bias}
\usage{
\method{print}{mbias}(x, ...)
}
\arguments{
\item{x}{An object of class 'mbias'.}

\item{...}{Other unused arguments.}
}
\value{
Print the observed and adjusted measures of association.
}
\description{
Print association corrected for M bias.
}

