% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probsens.R
\name{probsens}
\alias{probsens}
\title{Probabilistic sensitivity analysis.}
\usage{
probsens(
  case,
  exposed,
  type = c("exposure", "outcome"),
  reps = 1000,
  seca = list(dist = c("constant", "uniform", "triangular", "trapezoidal", "normal",
    "beta"), parms = NULL),
  seexp = NULL,
  spca = list(dist = c("constant", "uniform", "triangular", "trapezoidal", "normal",
    "beta"), parms = NULL),
  spexp = NULL,
  corr_se = NULL,
  corr_sp = NULL,
  alpha = 0.05
)
}
\arguments{
\item{case}{Outcome variable. If a variable, this variable is tabulated against.}

\item{exposed}{Exposure variable.}

\item{type}{Choice of correction for exposure or outcome misclassification.}

\item{reps}{Number of replications to run.}

\item{seca}{List defining sensitivity among cases:
\enumerate{
\item The sensitivity of exposure classification among those with the outcome
(when \code{type = "exposure"}), or
\item The sensitivity of outcome classification among those with the exposure
(when \code{type = "outcome"}).
}
The first argument provides the probability distribution function (constant,
uniform, triangular, trapezoidal, truncated normal, or beta) and the second
its parameters as a vector. Lower and upper bounds of the truncated normal
have to be between 0 and 1.
\enumerate{
\item constant: constant value,
\item uniform: min, max,
\item triangular: lower limit, upper limit, mode,
\item trapezoidal: min, lower mode, upper mode, max,
\item normal: lower bound, upper bound, mean, sd.
\item beta: alpha, beta.
}}

\item{seexp}{List defining sensitivity among controls:
\enumerate{
\item The sensitivity of exposure classification among those without the
outcome (when \code{type = "exposure"}), or
\item The sensitivity of outcome classification among those without the
exposure (when \code{type = "outcome"}).
}}

\item{spca}{List as above for \code{seca} but for specificity.}

\item{spexp}{List as above for \code{seexp} but for specificity.}

\item{corr_se}{Correlation between case and non-case sensitivities.}

\item{corr_sp}{Correlation between case and non-case specificities.}

\item{alpha}{Significance level.}
}
\value{
A list with elements:
\item{obs_data}{The analyzed 2 x 2 table from the observed data.}
\item{obs_measures}{A table of observed relative risk and odds ratio with
confidence intervals.}
\item{adj_measures}{A table of corrected relative risks and odds ratios.}
\item{sim_df}{Data frame of random parameters and computed values.}
\item{reps}{Number of replications.}
}
\description{
Summary-level probabilistic sensitivity analysis to correct for exposure
misclassification or outcome misclassification and random error.
Non-differential misclassification is assumed when only the two bias
parameters \code{seca} and \code{spca} are provided. Adding the 2
parameters \code{seexp} and \code{spexp} (i.e. providing the 4 bias
parameters) evaluates a differential misclassification.
}
\details{
Correlations between sensitivity (or specificity) of exposure classification
among cases and controls can be specified and use the NORmal To Anything
(NORTA) transformation (Li & Hammond, 1975).

In case of negative (<=0) adjusted count in the 2-by-2 table following given
prior Se/Sp distribution(s), draws are discarded.
}
\section{Updated calculations}{

episensr 2.0.0 introduced updated calculations of probabilistic bias analyses
by (1) using the NORTA transformation to define a correlation between
distributions, and (2) sampling true prevalences and then sampling the
adjusted cell counts rather than just using the expected cell counts from a
simple quantitative bias analysis. This updated version should be preferred
but if you need to run an old analysis, you can easily revert to the
computation using [probsens_legacy()] as follows:

```
library(episensr)
probsens <- probsens_legacy
```
}

\examples{
# The data for this example come from:
# Greenland S., Salvan A., Wegman D.H., Hallock M.F., Smith T.J.
# A case-control study of cancer mortality at a transformer-assembly facility.
# Int Arch Occup Environ Health 1994; 66(1):49-54.
greenland <- matrix(c(45, 94, 257, 945), dimnames = list(c("BC+", "BC-"),
c("Smoke+", "Smoke-")), nrow = 2, byrow = TRUE)
set.seed(123)
# Exposure misclassification, non-differential
probsens(greenland, type = "exposure", reps = 20000,
seca = list("trapezoidal", c(.75, .85, .95, 1)),
spca = list("trapezoidal", c(.75, .85, .95, 1)))

# Exposure misclassification, differential
probsens(greenland, type = "exposure", reps = 20000,
seca = list("trapezoidal", c(.75, .85, .95, 1)),
seexp = list("trapezoidal", c(.7, .8, .9, .95)),
spca = list("trapezoidal", c(.75, .85, .95, 1)),
spexp = list("trapezoidal", c(.7, .8, .9, .95)),
corr_se = .8,
corr_sp = .8)

probsens(greenland, type = "exposure", reps = 20000,
seca = list("beta", c(908, 16)),
seexp = list("beta", c(156, 56)),
spca = list("beta", c(153, 6)),
spexp = list("beta", c(205, 18)),
corr_se = .8,
corr_sp = .8)

probsens(matrix(c(338, 490, 17984, 32024),
dimnames = list(c("BC+", "BC-"), c("Smoke+", "Smoke-")), nrow = 2, byrow = TRUE),
type = "exposure",
reps = 1000,
seca = list("trapezoidal", c(.8, .9, .9, 1)),
spca = list("trapezoidal", c(.8, .9, .9, 1)))

# Disease misclassification
probsens(matrix(c(173, 602, 134, 663),
dimnames = list(c("BC+", "BC-"), c("Smoke+", "Smoke-")), nrow = 2, byrow = TRUE),
type = "outcome",
reps = 20000,
seca = list("uniform", c(.8, 1)),
spca = list("uniform", c(.8, 1)))

probsens(matrix(c(338, 490, 17984, 32024),
dimnames = list(c("BC+", "BC-"), c("Smoke+", "Smoke-")), nrow = 2, byrow = TRUE),
type = "outcome",
reps = 20000,
seca = list("uniform", c(.2, .6)),
seexp = list("uniform", c(.1, .5)),
spca = list("uniform", c(.99, 1)),
spexp = list("uniform", c(.99, 1)),
corr_se = .8,
corr_sp = .8)

probsens(matrix(c(173, 602, 134, 663),
dimnames = list(c("BC+", "BC-"), c("Smoke+", "Smoke-")), nrow = 2, byrow = TRUE),
type = "outcome",
reps = 20000,
seca = list("beta", c(100, 5)),
seexp = list("beta", c(110, 10)),
spca = list("beta", c(120, 15)),
spexp = list("beta", c(130, 30)),
corr_se = .8,
corr_sp = .8)

# Fox M.P., MacLehose R.F., Lash T.L.
# SAS and R code for probabilistic quantitative bias analysis for
# misclassified binary variables and binary unmeasured confounders
# Int J Epidemiol 2023:1624-1633.
fox <- matrix(c(40, 20, 60, 80),
dimnames = list(c("Diseased", "Non-diseased"), c("Exposed", "Unexposed")),
nrow = 2, byrow = TRUE)
set.seed(123)
probsens(fox, type = "exposure", reps = 10^6,
seca = list("beta", c(25, 3)),
spca = list("trapezoidal", c(.9, .93, .97, 1)),
seexp = list("beta", c(47, 7)),
spexp = list("trapezoidal", c(.8, .83, .87, .9)),
corr_se = .8,
corr_sp = .8)
}
\references{
Fox, M.P, MacLehose, R.F., Lash, T.L., 2021 \emph{Applying Quantitative
Bias Analysis to Epidemiologic Data}, pp.233--290, Springer.

Li, S.T., Hammond, J.L., 1975. \emph{Generation of Pseudorandom Numbers
with Specified Univariate Distributions and Correlation Coefficients}.
IEEE Trans Syst Man Cybern 5:557-561.
}
