% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probsens_pv.R
\name{probsens_pv}
\alias{probsens_pv}
\title{Probabilistic sensitivity analysis using predictive values.}
\usage{
probsens_pv(
  case,
  exposed,
  reps = 1000,
  ppvca = c(shape1, shape2),
  ppvexp = c(shape1, shape2),
  npvca = c(shape1, shape2),
  npvexp = c(shape1, shape2),
  alpha = 0.05
)
}
\arguments{
\item{case}{Outcome variable. If a variable, this variable is tabulated against.}

\item{exposed}{Exposure variable.}

\item{reps}{Number of replications to run.}

\item{ppvca}{Vector providing beta distribution parameters (alpha and beta)
for the positive predictive values among cases.}

\item{ppvexp}{Vector providing beta distribution parameters (alpha and beta)
for the positive predictive values among controls.}

\item{npvca}{Vector as above for \code{ppvca} but for negative predictive value.}

\item{npvexp}{Vector as above for \code{ppvexp} but for negative predictive value.}

\item{alpha}{Significance level.}
}
\value{
A list with elements:
\item{obs_data}{The analyzed 2 x 2 table from the observed data.}
\item{obs_measures}{A table of observed relative risk and odds ratio with
confidence intervals.}
\item{adj_measures}{A table of corrected relative risks and odds ratios.}
\item{sim_df}{Data frame of random parameters and computed values.}
\item{reps}{Number of replications.}
}
\description{
Probabilistic sensitivity analysis to correct for exposure misclassification
or outcome misclassification and random error, using predictive values.
}
\examples{
# The data for this example come from:
# Greenland S., Salvan A., Wegman D.H., Hallock M.F., Smith T.J.
# A case-control study of cancer mortality at a transformer-assembly facility.
# Int Arch Occup Environ Health 1994; 66(1):49-54.
#greenland <- matrix(c(45, 94, 257, 945), dimnames = list(c("BC+", "BC-"),
#c("Smoke+", "Smoke-")), nrow = 2, byrow = TRUE)
#set.seed(123)
# Exposure misclassification, non-differential
#probsens(greenland, type = "exposure", reps = 20000,
#seca = list("trapezoidal", c(.75, .85, .95, 1)),
#spca = list("trapezoidal", c(.75, .85, .95, 1)))
}
\references{
Fox, M.P, MacLehose, R.F, Lash, T.L., 2021 \emph{Applying Quantitative
Bias Analysis to Epidemiologic Data}, pp.253--255, Springer.
}
