\name{episensr-package}
\alias{episensr-package}
\alias{episensr}
\docType{package}
\title{
Basic sensitivity analysis of epidemiological data
}
\description{
episensr provides basic sensitivity analysis of the observed relative
risk adjusting for unmeasured confounding and misclassification of the exposure.
}
\details{
\tabular{ll}{
Package: \tab episensr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-06\cr
License: \tab GPL-2\cr
}
}
\author{
Denis Haine \email{denis.haine@gmail.com}
}
\references{Lash, T.L., Fox, M.P, Fink, A.K., 2009 \emph{Applying
    Quantitative Bias Analysis to Epidemiologic Data}, pp.43--58, Springer. 
}
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
