\name{episensr-package}
\alias{episensr-package}
\alias{episensr}
\docType{package}
\title{episensr: Basic sensitivity analysis of epidemiological results}
\description{
  'episensr' provides basic sensitivity analysis of the observed relative
  risks adjusting for unmeasured confounding and misclassification of
  the exposure/outcome, or both.}
\author{
Maintainer: Denis Haine \email{denis.haine@gmail.com}
}
\references{
  Lash, T.L., Fox, M.P, Fink, A.K., 2009 \emph{Applying Quantitative
    Bias Analysis to Epidemiologic Data}, Springer.
}
\seealso{
  Useful links:
  \itemize{
    \item \url{https://github.com/dhaine/episensr}
    \item Report bugs at \url{https://github.com/dhaine/episensr/issues}
  }
}
