\name{episensr}
\alias{episensr}
\alias{episensr-package}
\docType{package}
\title{Basic sensitivity analysis of epidemiological results}
\description{
  episensr provides basic sensitivity analysis of the observed relative
  risks adjusting for unmeasured confounding and misclassification of
  the exposure/outcome, or both.
}
\details{
  \tabular{ll}{
    Package: \tab episensr\cr
    Type: \tab Package\cr
    Version: \tab 0.6\cr
    Date: \tab 2015-03-24\cr
    License: \tab GPL-2\cr
  }
}
\author{
  Denis Haine \email{denis.haine@gmail.com}
}
\references{
  Lash, T.L., Fox, M.P, Fink, A.K., 2009 \emph{Applying Quantitative
    Bias Analysis to Epidemiologic Data}, Springer.
}
