% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.episensr.R
\name{print.episensr}
\alias{print.episensr}
\title{Print associations for episensr class}
\usage{
\method{print}{episensr}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{An object of class 'episensr'.}

\item{digits}{Minimal number of _significant_ digits, see 'print.default'.}

\item{...}{Other unused arguments.}
}
\value{
Print the observed and adjusted measures of association.
}
\description{
Print associations for episensr objects.
}

