% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probsens.conf.R
\name{probsens.conf}
\alias{probsens.conf}
\title{Probabilistic sensitivity analysis for unmeasured confounding.}
\usage{
probsens.conf(
  case,
  exposed,
  reps = 1000,
  prev_exp = list(dist = c("constant", "uniform", "triangular", "trapezoidal", "normal",
    "beta"), parms = NULL),
  prev_nexp = list(dist = c("constant", "uniform", "triangular", "trapezoidal", "normal",
    "beta"), parms = NULL),
  risk = list(dist = c("constant", "uniform", "triangular", "trapezoidal",
    "log-logistic", "log-normal"), parms = NULL),
  corr_p = NULL,
  discard = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{case}{Outcome variable. If a variable, this variable is tabulated against.}

\item{exposed}{Exposure variable.}

\item{reps}{Number of replications to run.}

\item{prev_exp}{List defining the prevalence of exposure among the exposed.
The first argument provides the probability distribution function (constant,
uniform, triangular, trapezoidal, truncated normal, or beta) and the second its
parameters as a vector. Lower bound of the truncated normal cannot be less than
zero. Upper bound is Inf by default.
\enumerate{
\item constant: constant value,
\item uniform: min, max,
\item triangular: lower limit, upper limit, mode,
\item trapezoidal: min, lower mode, upper mode, max.
\item normal: lower bound, upper bound, mean, sd.
\item beta: alpha, beta.
}}

\item{prev_nexp}{List defining the prevalence of exposure among the unexposed.}

\item{risk}{List defining the confounder-disease relative risk or the confounder-exposure
odds ratio. The first argument provides the probability distribution function
(constant, uniform, triangular, trapezoidal, log-logistic, or log-normal) and
the second its parameters as a vector:
\enumerate{
\item constant: constant value,
\item uniform: min, max,
\item triangular: lower limit, upper limit, mode,
\item trapezoidal: min, lower mode, upper mode, max.
\item log-logistic: shape, rate. Must be strictly positive,
\item log-normal: meanlog, sdlog. This is the mean and standard deviation on the log scale.
}}

\item{corr_p}{Correlation between the exposure-specific confounder prevalences.}

\item{discard}{A logical scalar. In case of negative adjusted count, should
the draws be discarded? If set to FALSE, negative counts are set to zero.}

\item{alpha}{Significance level.}
}
\value{
A list with elements:
\item{obs_data}{The analyzed 2 x 2 table from the observed data.}
\item{obs_measures}{A table of observed relative risk and odds ratio with confidence intervals.}
\item{adj_measures}{A table of corrected relative risks and odds ratios.}
\item{sim_df}{Data frame of random parameters and computed values.}
\item{reps}{Number of replications.}
}
\description{
Probabilistic sensitivity analysis to correct for unknown or unmeasured
confounding and random error simultaneously.
}
\details{
Correlations between prevalences of exposure classification among cases and
controls can be specified and use the NORmal To Anything (NORTA)
transformation (Li & Hammond, 1975).
}
\section{Updated calculations}{

episensr 2.0.0 introduced updated calculations of probabilistic bias analyses
by (1) using the NORTA transformation to define a correlation between
distributions, and (2) sampling true prevalences and then sampling the
adjusted cell counts rather than just using the expected cell counts from a
simple quantitative bias analysis. This updated version should be preferred
but if you need to run an old analysis, you can easily revert to the
computation using [probsens.conf_legacy()] as follows:

```
library(episensr)
probsens.conf <- probsens.conf_legacy
```
}

\examples{
# The data for this example come from:
# Tyndall M.W., Ronald A.R., Agoki E., Malisa W., Bwayo J.J., Ndinya-Achola J.O. et al.
# Increased risk of infection with human immunodeficiency virus type 1 among
# uncircumcised men presenting with genital ulcer disease in Kenya.
# Clin Infect Dis 1996;23:449-53.
tyndall <- matrix(c(105, 85, 527, 93),
dimnames = list(c("HIV+", "HIV-"), c("Circ+", "Circ-")), nrow = 2, byrow = TRUE)
set.seed(123)
probsens.conf(tyndall, reps = 20000,
prev_exp = list("triangular", c(.7, .9, .8)),
prev_nexp = list("trapezoidal", c(.03, .04, .05, .06)),
risk = list("triangular", c(.6, .7, .63)),
corr_p = .8)

set.seed(123)
probsens.conf(tyndall, reps = 20000,
prev_exp = list("beta", c(200, 56)),
prev_nexp = list("beta", c(10, 16)),
risk = list("triangular", c(.6, .7, .63)),
corr_p = .8)

set.seed(123)
probsens.conf(tyndall, reps = 20000,
prev_exp = list("normal", c(.01, .12, 0.03, 0.005)),
prev_nexp = list("normal", c(0, Inf, 0.01, 0.0001)),
risk = list("triangular", c(.6, .7, .63)), corr_p = .8)
}
\references{
Lash, T.L., Fox, M.P, Fink, A.K., 2009 \emph{Applying Quantitative
Bias Analysis to Epidemiologic Data}, pp.117--150, Springer.

Li, S.T., Hammond, J.L., 1975. \emph{Generation of Pseudorandom Numbers
with Specified Univariate Distributions and Correlation Coefficients}.
IEEE Trans Syst Man Cybern 5:557-561.
}
