% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probsens.irr.R
\name{probsens.irr}
\alias{probsens.irr}
\title{Probabilistic sensitivity analysis for exposure misclassification of person-time data and random error.}
\usage{
probsens.irr(counts, pt = NULL, reps = 1000, seca.parms = list(dist =
  c("constant", "uniform", "triangular", "trapezoidal", "logit-logistic",
  "logit-normal"), parms = NULL), seexp.parms = NULL, spca.parms = list(dist
  = c("constant", "uniform", "triangular", "trapezoidal", "logit-logistic",
  "logit-normal"), parms = NULL), spexp.parms = NULL, corr.se = NULL,
  corr.sp = NULL, discard = TRUE, alpha = 0.05, dec = 4, print = TRUE)
}
\arguments{
\item{counts}{A table or matrix where first row contains disease counts and second row contains person-time at risk, and first and second columns are exposed and unexposed observations, as:
\tabular{lll}{
\tab Exposed \tab Unexposed \cr
Cases \tab a \tab b \cr
Person-time \tab N1 \tab N0
}}

\item{pt}{A numeric vector of person-time at risk. If provided, \code{counts} must be a numeric vector of disease counts.}

\item{reps}{Number of replications to run.}

\item{seca.parms}{List defining the sensitivity of exposure classification among those with the outcome. The first argument provides the probability distribution function (uniform, triangular, trapezoidal, logit-logistic, or logit-normal) and the second its parameters as a vector:
\enumerate{
\item Constant: constant value,
\item Uniform: min, max,
\item Triangular: lower limit, upper limit, mode,
\item Trapezoidal: min, lower mode, upper mode, max,
\item Logit-logistic: location, scale, lower bound shift, upper bound shift,
\item Logit-normal: location, scale, lower bound shift, upper bound shift.
}}

\item{seexp.parms}{List defining the sensitivity of exposure classification among those without the outcome.}

\item{spca.parms}{List defining the specificity of exposure classification among those with the outcome.}

\item{spexp.parms}{List defining the specifity of exposure classification among those without the outcome.}

\item{corr.se}{Correlation between case and non-case sensitivities.}

\item{corr.sp}{Correlation between case and non-case specificities.}

\item{discard}{A logical scalar. In case of negative adjusted count, should the draws be discarded? If set to FALSE, negative counts are set to zero.}

\item{alpha}{Significance level.}

\item{dec}{Number of decimals in the printout.}

\item{print}{A logical scalar. Should the results be printed?}
}
\value{
A list with elements:
\item{obs.data}{The analysed 2 x 2 table from the observed data.}
\item{obs.measures}{A table of observed incidence rate ratio with exact confidence interval.}
\item{adj.measures}{A table of corrected incidence rate ratios.}
\item{sim.df}{Data frame of random parameters and computed values.}
}
\description{
Probabilistic sensitivity analysis to correct for exposure misclassification when person-time data has been collected.
}
\examples{
set.seed(123)
# Exposure misclassification, non-differential
probsens.irr(matrix(c(2, 67232, 58, 10539000),
dimnames = list(c("GBS+", "Person-time"), c("HPV+", "HPV-")), ncol = 2),
reps = 20000,
seca.parms = list("trapezoidal", c(.4, .45, .55, .6)),
spca.parms = list("constant", 1))
}
\references{
Lash, T.L., Fox, M.P, Fink, A.K., 2009 \emph{Applying Quantitative
Bias Analysis to Epidemiologic Data}, pp.117--150, Springer.
}

