% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{selection}
\alias{selection}
\title{Sensitivity analysis to correct for selection bias.}
\usage{
selection(case, exposed, bias_parms = NULL, alpha = 0.05)
}
\arguments{
\item{case}{Outcome variable. If a variable, this variable is tabulated
against.}

\item{exposed}{Exposure variable.}

\item{bias_parms}{Selection probabilities. Either a vector of 4 elements
between 0 and 1 defining the following probabilities in this order can be
provided:
\enumerate{
\item Selection probability among cases exposed (1),
\item Selection probability among cases unexposed (2),
\item Selection probability among noncases exposed (3), and
\item Selection probability among noncases unexposed (4).
}
or a single positive selection-bias factor which is the ratio of the exposed
versus unexposed selection probabilities comparing cases and noncases [(1*4)/(2*3)
from above].}

\item{alpha}{Significance level.}
}
\value{
A list with elements:
\item{model}{Bias analysis performed.}
\item{obs.data}{The analyzed 2 x 2 table from the observed data.}
\item{corr.data}{The same table corrected for  selection proportions.}
\item{obs.measures}{A table of odds ratios and relative risk with confidence intervals.}
\item{adj.measures}{Selection bias corrected measures of outcome-exposure relationship.}
\item{bias.parms}{Input bias parameters: selection probabilities.}
\item{selbias.or}{Selection bias odds ratio based on the bias parameters chosen.}
}
\description{
Simple sensitivity analysis to correct for selection bias using estimates of
the selection proportions.
}
\examples{
# The data for this example come from:
# Stang A., Schmidt-Pokrzywniak A., Lehnert M., Parkin D.M., Ferlay J., Bornfeld N.
# et al.
# Population-based incidence estimates of uveal melanoma in Germany. Supplementing
# cancer registry data by case-control data.
# Eur J Cancer Prev 2006;15:165-70.
selection(matrix(c(136, 107, 297, 165),
dimnames = list(c("UM+", "UM-"), c("Mobile+", "Mobile-")),
nrow = 2, byrow = TRUE),
bias_parms = c(.94, .85, .64, .25))


selection(matrix(c(136, 107, 297, 165),
dimnames = list(c("UM+", "UM-"), c("Mobile+", "Mobile-")),
nrow = 2, byrow = TRUE),
bias_parms = 0.43)
}
